Move api modules into a dedicated directory
