+++ frontend/src/api/index.ts
@@ -1,4 +1,4 @@
-import session from "./session";
+import session from "./modules/session";
 
 export default {
   session,
