+++ frontend/src/api/modules/session.ts
@@ -1,6 +1,6 @@
 import type { AxiosResponse } from "axios";
-import type { loginForm, User, UserWithPassword } from "./types";
-import instance from "./instance";
+import type { loginForm, User, UserWithPassword } from "../types";
+import instance from "../instance";
 
 export default {
   login: (form: loginForm): Promise<AxiosResponse<User>> => {
